重构: 按 Bell 6.X 契约重写 Solon Xbatis 后端

+++ jimuqu-modules/jimuqu-system/src/main/java/com/jimuqu/system/domain/vo/CacheListInfoVo.java
@@ -16,7 +16,7 @@ public class CacheListInfoVo {
 
     private Properties info;
 
-    private Long dbSize;
+    private Integer dbSize;
 
     private List<Map<String, String>> commandStats;
